refactor to use npm workspaces
