+++ packages/ace-linters/components/description-tooltip.ts
