+++ packages/ace-linters/language-provider.ts
@@ -2,10 +2,11 @@ import {Ace, Range as AceRange} from "ace-code";
 import {MessageController} from "./message-controller";
 import {DescriptionTooltip} from "./components/description-tooltip";
 import Tooltip = AceLinters.Tooltip;
-import {AceLinters} from "./services/language-service";
+import {AceLinters} from "./index";
 import TextEdit = AceLinters.TextEdit;
 import {FormattingOptions} from "vscode-languageserver-types";
 import {CommonConverter} from "./type-converters/common-converters";
+import {MarkDownConverter} from "./type-converters/converters";
 
 let showdown = require('showdown');
