+++ packages/ace-linters/services/typescript/lib/typescriptServices.d.ts
