+++ packages/ace-linters/type-converters/converters.d.ts
@@ -1,3 +1,3 @@
-interface MarkDownConverter {
+export interface MarkDownConverter {
     makeHtml(markdownText: string): string;
 }
