+++ packages/demo/docs-example/typescript-example.js
