+++ packages/demo/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "@ace-linters/demo"
+}
